refactor(auth): migrate to standardized mapper pattern - complete auth module architectural alignment
## Core Changes

### 🗂️ Mapper Introduction
- **NEW**: `supabaseAuthMapper.ts` - Centralized domain-database mapping functions
  - `mapSupabaseUserToAuthUser()` - User entity transformation
  - `mapSupabaseSessionToAuthSession()` - Session data mapping
  - Exported as `supabaseAuthMapper` object for consistent access pattern

### 🏗️ Directory Structure Standardization
- **MOVED**: `supabaseAuthRepository.ts` → `/supabase/supabaseAuthRepository.ts`
- **ENHANCED**: Repository implementation with proper error handling
- **UPDATED**: Import paths in application layer and tests

### 🔧 Repository Implementation Improvements
- **ERROR HANDLING**: Added `createErrorHandler('infrastructure', 'Auth')` pattern
- **MAPPER INTEGRATION**: Repository now uses centralized mapper functions
- **TYPE SAFETY**: Proper error type handling without assertions
- **DEBUGGING**: Maintained existing debug functionality

### 📝 Import Path Updates
- **APPLICATION**: Updated `auth.ts` to use new repository path
- **TESTS**: Updated `auth.test.ts` mock path to match new structure

## Technical Benefits

### ✅ Architectural Consistency
- Auth module now follows identical pattern as all other modules
- Centralized mapping logic for better maintainability
- Clear separation between domain transformation and repository logic

### ✅ Error Handling Enhancement
- Standardized error handling across all auth operations
- Proper error context for debugging and monitoring
- Type-safe error handling without TypeScript assertions

### ✅ Testability Improvements
- Mapper functions can be unit tested independently
- Repository logic separated from transformation concerns
- Mock compatibility maintained for existing tests

## Files Modified: 4 files
- ✅ Auth mapper: Created standardized mapping functions
- ✅ Auth repository: Enhanced with error handling and mapper integration
- ✅ Application layer: Updated import paths
- ✅ Test layer: Updated mock paths

Auth module now fully aligned with established architectural patterns.
